// src/sharedmodule/interface/Interface.tsx
export interface QuestionOptions {
  A: string;
  B: string;
  C: string;
  D: string;
}
// ...
export interface QuestionListOfProps {
     "_id":string,
        "title": string,
        "description": string,
        "options":keyof QuestionOptions,
        "answer": string,
        "status": "active"| "inactive",
        "instructor": string,
        "difficulty": "easy"|  "medium" | "hard",
        "points": number,
        "type": "FE"| "BE"| "DB",
}

export interface CreateQuestionPayload {
  title: string;
  description: string;
  options: {
    A: string;
    B: string;
    C: string;
    D: string;
  };
  answer: "A" | "B" | "C" | "D";
  difficulty: "easy" | "medium" | "hard";
  type: "BE" | "FE";
}